sparkles: feat: Validações de positions name

+++ SeniorCareManager.WebAPI/Controllers/PositionController.cs
@@ -1,7 +1,11 @@
 ﻿using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
+using SeniorCareManager.WebAPI.Objects.Dtos.Entities;
+using SeniorCareManager.WebAPI.Objects.Enums;
 using SeniorCareManager.WebAPI.Objects.Models;
 using SeniorCareManager.WebAPI.Services.Interfaces;
+using SeniorCareManager.WebAPI.Services.Utils;
+using SeniorCareManager.WebAPI.Objects.Contracts;
 
 namespace SeniorCareManager.WebAPI.Controllers;
 
@@ -10,54 +14,123 @@ namespace SeniorCareManager.WebAPI.Controllers;
 public class PositionController: Controller
 {
     private readonly IPositionService _positionService;
+    private readonly Response _response;
 
     public PositionController(IPositionService service)
     {
         this._positionService = service;
+        _response = new Response();
     }
 
     [HttpGet]
     public async Task<IActionResult> Get()
     {
         var position = await _positionService.GetAll();
-        return Ok(position);
+        _response.Code = ResponseEnum.Success;
+        _response.Data = position;
+        _response.Message = "Lista de cargos!";
+        return Ok(_response);
     }
 
     [HttpGet("{id}")]
     public async Task<IActionResult> GetById(int id)
     {
-        var position = await _positionService.GetById(id);
-        if (position == null) return NotFound("Cargo não encontrado!");
-        return Ok(position);
+        try
+        {
+            var position = await _positionService.GetById(id);
+            if (position is null)
+            {
+                _response.Code = ResponseEnum.NotFound;
+                _response.Message = "Cargo não encontrado.";
+                _response.Data = position;
+                return NotFound(_response);
+            }
+
+            _response.Code = ResponseEnum.Success;
+            _response.Message = "Cargo " + position.Name + " obtido com sucesso!";
+            _response.Data = position;
+            return Ok(_response);
+
+        }
+        catch (Exception ex)
+        {
+            _response.Code = ResponseEnum.NotFound;
+            _response.Message = "Não foi possível adquirir o cargo.";
+            _response.Data = null;
+            return StatusCode(StatusCodes.Status500InternalServerError, _response);
+        }
     }
 
     [HttpPost]
-    public async Task<IActionResult> Post(Position position)
+    public async Task<IActionResult> Post(PositionDTO positionDto)
     {
+        if (string.IsNullOrWhiteSpace(positionDto.Name))
+        {
+            _response.Code = ResponseEnum.Invalid;
+            _response.Data = null;
+            _response.Message = "Dados inválidos.";
+
+            return BadRequest(_response);
+        }
+        var positions = await _positionService.GetAll();
+        if (CheckDuplicates(positions, positionDto))
+        {
+            _response.Code = ResponseEnum.Conflict;
+            _response.Data = positionDto;
+            _response.Message = "Nome duplicado.";
+            return BadRequest(_response);
+        }
         try
         {
-            await _positionService.Create(position);
+            positionDto.Id = 0;
+            await _positionService.Create(positionDto);
+            _response.Code = ResponseEnum.Success;
+            _response.Message = "Cargo Cadastrado com sucesso!";
+            _response.Data = positionDto;
         }
         catch (Exception ex)
         {
-            return StatusCode(500, "Ocorreu um erro ao tentar inserir um novo cargo.");
+            _response.Code = ResponseEnum.Error;
+            _response.Message = "Não foi possível cadastrar o cargo.";
+            _response.Data = positionDto;
+            return StatusCode(StatusCodes.Status500InternalServerError, _response);
         }
-        return Ok(position);
+        return Ok(_response);
     }
 
     [HttpPut("{id}")]
-    public async Task<IActionResult> Put(int id, Position position)
+    public async Task<IActionResult> Put(int id, PositionDTO positionDto)
     {
+        if (string.IsNullOrWhiteSpace(positionDto.Name))
+        {
+            _response.Code = ResponseEnum.Invalid;
+            _response.Data = positionDto;
+            _response.Message = "Nome inválido.";
+            return BadRequest(_response);
+        }
+        var positions = await _positionService.GetAll();
+        if (!CheckDuplicates(positions, positionDto))
+        {
+            _response.Code = ResponseEnum.Conflict;
+            _response.Data = positionDto;
+            _response.Message = "Nome duplicado.";
+            return BadRequest(_response);
+        }
         try
         {
-            await _positionService.Update(position, id);
+            await _positionService.Update(positionDto, id); ;
+            _response.Code = ResponseEnum.Success;
+            _response.Message = "Cargo alterado com sucesso!";
+            _response.Data = positionDto;
         }
         catch (Exception ex)
         {
-            return StatusCode(500, "Ocorreu um erro ao tentar atualizar o cargo: " + ex.Message);
+            _response.Code = ResponseEnum.Error;
+            _response.Message = "Não foi possível alterar o cargo!";
+            _response.Data = positionDto;
+            return StatusCode(StatusCodes.Status500InternalServerError, _response);
         }
-
-        return Ok(position);
+        return Ok(_response);
     }
 
     [HttpDelete("{id}")]
@@ -65,14 +138,44 @@ public class PositionController: Controller
     {
         try
         {
+            var position = await _positionService.GetById(id);
+            if (position is null)
+            {
+                _response.Code = ResponseEnum.NotFound;
+                _response.Data = null;
+                _response.Message = "O cargo não foi encontrado.";
+                return NotFound(_response);
+            }
             await _positionService.Remove(id);
-        }
-        catch (Exception ex)
-        {
-            return StatusCode(500, "Ocorreu um erro ao tentar remover o cargo.");
+            _response.Code = ResponseEnum.Success;
+            _response.Message = "Grupo de cargo apagado com sucesso!";
+            _response.Data = null;
         }
 
-        return Ok("cargo apagado com sucesso");
+        catch (Exception ex)
+        {
+            _response.Code = ResponseEnum.Error;
+            _response.Message = "Erro ao tentar apagar grupo de cargo.";
+            _response.Data = null;
+        }
+        return Ok(_response);
+
+    }
+    private static bool CheckDuplicates(IEnumerable<PositionDTO> positionsDTO, PositionDTO positionDTO)
+    {
+        foreach (var position in positionsDTO)
+        {
+            if (positionDTO.Id == position.Id)
+            {
+                continue;
+            }
+
+            if (StringValidator.CompareString(positionDTO.Name, position.Name))
+            {
+                return true;
+            }
+        }
+        return false;
     }
